separated php8 only tests
